pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "layout-lens"
version = "0.1.0"
description = "An agentic prompt to 3D placeholder visualizer (LayoutLens)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic-ai-slim[openai]",
  "pydantic-settings",
  "python-dotenv",
  "shapely",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
]

[project.scripts]
layout-lens = "layout_lens.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"